requirements.txt
# python 3.10
polygon-api-client
polars==0.18.7